Move GroupedItem to common types
